# 20250201_demo_webapp/backend/server/models/post.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from server.models.base_model import Model
from server import util


class Post(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, user_id=None, content=None, created_at=None):  # noqa: E501
        """Post - a model defined in OpenAPI

        :param id: The id of this Post.  # noqa: E501
        :type id: int
        :param user_id: The user_id of this Post.  # noqa: E501
        :type user_id: int
        :param content: The content of this Post.  # noqa: E501
        :type content: str
        :param created_at: The created_at of this Post.  # noqa: E501
        :type created_at: datetime
        """
        self.openapi_types = {
            'id': int,
            'user_id': int,
            'content': str,
            'created_at': datetime
        }

        self.attribute_map = {
            'id': 'id',
            'user_id': 'userId',
            'content': 'content',
            'created_at': 'createdAt'
        }

        self._id = id
        self._user_id = user_id
        self._content = content
        self._created_at = created_at

    @classmethod
    def from_dict(cls, dikt) -> 'Post':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Post of this Post.  # noqa: E501
        :rtype: Post
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Post.


        :return: The id of this Post.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Post.


        :param id: The id of this Post.
        :type id: int
        """

        self._id = id

    @property
    def user_id(self) -> int:
        """Gets the user_id of this Post.


        :return: The user_id of this Post.
        :rtype: int
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: int):
        """Sets the user_id of this Post.


        :param user_id: The user_id of this Post.
        :type user_id: int
        """

        self._user_id = user_id

    @property
    def content(self) -> str:
        """Gets the content of this Post.


        :return: The content of this Post.
        :rtype: str
        """
        return self._content

    @content.setter
    def content(self, content: str):
        """Sets the content of this Post.


        :param content: The content of this Post.
        :type content: str
        """

        self._content = content

    @property
    def created_at(self) -> datetime:
        """Gets the created_at of this Post.


        :return: The created_at of this Post.
        :rtype: datetime
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at: datetime):
        """Sets the created_at of this Post.


        :param created_at: The created_at of this Post.
        :type created_at: datetime
        """

        self._created_at = created_at